requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
requests==2.32.3
aiohttp==3.9.5
loguru==0.7.2
numpy==2.1.2
fuzzywuzzy==0.18.0
PyYAML==6.0.2